+++ packages/actions/src/actions/index.ts
@@ -4,6 +4,7 @@ import http  from './http';
 import pickData from "./pkData";
 import table from "./table";
 import format from './format'
+import lib from './lib';
 import {IAction} from "../typing/IAction";
 
 
@@ -13,27 +14,31 @@ export {
     pickData,
     table,
     format,
-    getAction,
+}
+
+const actionMap: Record<string, IAction> = {
+    'pagenote/lib@v1': lib,
 }
 
 /**
  * 根据action 名称，获取执行方法
  * */
-function getAction(uses: string):Promise<IAction> {
-    const [_namespace = '', action] = uses.split('/');
+export function getAction(uses: string):Promise<IAction|null> {
+    const [_namespace = '', action=''] = uses.split('/');
     const [actionName = ''] = action.split('@');
+    if(actionMap[uses]){
+        return Promise.resolve(actionMap[uses]);
+    }
     switch (actionName) {
         case 'handlebars':
             return Promise.resolve(handlebars);
         case 'http':
             return Promise.resolve(http);
-        case 'pickData':
+        case 'pick':
             return Promise.resolve(pickData);
-        case 'table':
-            return Promise.resolve(table);
         case 'format':
             return Promise.resolve(format);
         default:
-            return Promise.reject(`未找到${uses}的实现`);
+            return Promise.resolve(null);
     }
 }