+++ packages/actions/src/runner/Background.ts
@@ -4,6 +4,7 @@ import set from 'lodash/set'
 import get from 'lodash/get'
 import {generateMatrixTasks, replaceTemplates} from "../utils";
 import {exprEval, ifCheck} from "../utils/expr-eval";
+import {getAction} from "../actions";
 
 /**
  * 适用于后台运行的运行器（service worker or nodejs，不依赖 DOM 结构）
@@ -30,18 +31,19 @@ export default class Background{
     [key:string]: any,
   }>();
 
+  public log: Map<string, any> = new Map<string, any>();
+
   /**运行期的配置信息*/
   private readonly  option: WorkflowOption;
 
-  public running: [number,number] = [-1,-1]
 
-  /** 初始化*/
   constructor(option:WorkflowOption) {
     this.option = option;
-    this._updateYml(option?.yml || '');
+    this.updateYml(option?.yml || '');
     this.context = {
       env:{},
       trigger:{},
+    //   todo 系统变量，如当前时间获取
     }
 
     this.state = WorkFlowState.waiting
@@ -51,9 +53,9 @@ export default class Background{
    * yml to object
    * https://nodeca.github.io/js-yaml/
    * */
-  _updateYml(yml?: string){
+  updateYml(yml: string | WorkFlow){
     try{
-      this.workflowInfo = yml ? jsYaml.load(yml) as WorkFlow : null;
+      this.workflowInfo = typeof yml === 'string' ? jsYaml.load(yml) as WorkFlow : yml;
     }catch (e) {
       throw e
     }
@@ -66,6 +68,7 @@ export default class Background{
   _setRuntime(id: string, value: any = {},contextId: string){
     const runtime = this.runtime.get(contextId) || this.runtime.set(contextId,{}).get(contextId) || {}
     set(runtime, id, value)
+    this.log.set(id,value);
   }
 
   /**
@@ -78,143 +81,13 @@ export default class Background{
   /**
    * 获取 job/step 下可被读取的上下文内容
    * */
-  getContextVariables(contextId: string){
-    const variables = {
+  _getContextVariables(contextId: string){
+    return {
       env: this.context.env,
       steps: this._getRuntime(contextId)?.steps || {},
       jobs: this._getRuntime(contextId)?.jobs || {},
       matrix: this._getRuntime(contextId)?.matrix || {}
     }
-    return variables
-  }
-  /**
-   * 执行单个 step 任务
-   * */
-  async runStep(step: Step,stepsContextId:string){
-    if(!step){
-      return
-    }
-    const { uses='' } = step;
-
-    step._state = TaskState.running;
-
-    const variables = this.getContextVariables(stepsContextId)
-
-
-    if(step.debug){
-      debugger
-    }
-    let response;
-
-    /**条件检测**/
-    if(step.if){
-      const checkResult = ifCheck(step.if,variables);
-      if(!checkResult){
-        step._state = TaskState.skip;
-      }
-    } else{
-      const withArgs = step.with ? replaceTemplates(step.with,variables) : undefined
-      try{
-        if(this.option.hooks?.beforeStep){
-          this.option.hooks?.beforeStep(step,withArgs);
-        }
-        const action = await this.option.registerAction(uses || '');
-        if(uses && !action){
-          step._state = TaskState.fail;
-          throw Error('without action::'+ uses)
-        }
-
-        if(step.run && step.run.length){
-          step.run.forEach(function (terminate) {
-            response = exprEval(terminate,variables)
-          })
-        }else{
-          try{
-            response = action ? await action(withArgs) : withArgs;
-          }catch (e) {
-            step._state = TaskState.fail;
-          }
-        }
-
-        this._setRuntime(`steps.${step.id || step.name}.outputs`, response,stepsContextId)
-      }catch (e) {
-        step._state = TaskState.fail;
-        throw e;
-      }
-      step._state = TaskState.complete;
-      if(this.option.hooks?.afterStep){
-        this.option.hooks?.afterStep(step,response);
-      }
-    }
-
-    /**由用户自行控制step 退出码*/
-    let exitCode: number | boolean = 0;
-    if(step.exit){
-      if(typeof step.exit === 'number'){
-        exitCode = step.exit;
-      }else{
-        exitCode = exprEval(step.exit,this.getContextVariables(stepsContextId))
-      }
-
-      if(exitCode!==0){
-        step._state = TaskState.fail;
-        throw Error((step.id || step.name) + ' exit code is not equal 0 :' + exitCode);
-      }
-    }
-    return response;
-  }
-
-  /**
-   * 执行多个任务组；用于控制并发、串行、失败控制等
-   * */
-  async _runSteps(steps: Step[],contextId:string){
-    let response;
-    for (let i = 0; i < steps.length; i++) {
-      try {
-        response = await this.runStep(steps[i],contextId);
-      }catch (e) {
-        if(steps[i]['continue-on-error']){
-          console.warn('允许失败，继续执行',e)
-        }else{
-          throw e;
-        }
-      }
-    }
-    return response;
-  }
-
-  /**
-   * 执行 定义的单个 job
-   * */
-  async runJob(job: Job,jobsContextId: string){
-    if(!job) return;
-    let jobResponse;
-    job._state = TaskState.running;
-    /**
-     * 如果 job 是一个 matrix 任务，则要进行叉乘创建子 job。
-     */
-    const variables = this._getRuntime(jobsContextId) || {};
-    job = replaceTemplates<Job>(job,variables)
-    if(job.strategy?.matrix){
-      const matrixVariables = generateMatrixTasks(job.strategy.matrix) || [];
-      for (let i=0; i<matrixVariables.length; i++){
-        for(let j=0; j<matrixVariables[i].length; j++){
-          const contextId = Date.now()+ (job.id || job.name) + 'matrix'+i+j;
-          const stepVariable = matrixVariables[i][j];
-          this._setRuntime('matrix',stepVariable,contextId)
-          jobResponse = await this._runSteps(job.steps,contextId);
-          this._setRuntime(`jobs.${job.id||job.name}.outputs`,jobResponse,jobsContextId)
-        }
-      }
-    } else {
-      const {id='',name='',} = job || {};
-      const contextId = id + name + Date.now();
-      jobResponse = await this._runSteps(job.steps,contextId)
-      this._setRuntime(`jobs.${id||name}.outputs`,jobResponse,jobsContextId)
-    }
-
-    job._state = TaskState.complete;
-    return jobResponse;
   }
 
   /** 开始前的 环境变量准备*/
@@ -235,6 +108,23 @@ export default class Background{
   }
 
   /**
+   * 配置的测试用例，用于反馈配置是否合法，避免在正式运行时出现一些错误
+   * */
+  async runTest(){
+    await this._prepareEnv();
+    const testJobs = this.workflowInfo?.test || [];
+
+    for(let i=0; i<testJobs.length; i++){
+      await this._runJob(testJobs[i],`test${Date.now()}`)
+    }
+
+    return testJobs.every(function (job) {
+      return job._state !== TaskState.fail
+    })
+  }
+
+
+  /**
    * 运行 workflow
    * */
   async run(){
@@ -246,28 +136,149 @@ export default class Background{
     /**所有的 jobs 同一个上下文*/
     const contextId = Date.now()+ (this.workflowInfo?.name || '')
     for(let i=0; i<jobs.length; i++){
-      await this.runJob(jobs[i],contextId)
+      await this._runJob(jobs[i],contextId)
     }
-    this.state = WorkFlowState.success;
+    this.state = WorkFlowState.finished;
   }
 
   /**
-   * 配置的测试用例，用于反馈配置是否合法，避免在正式运行时出现一些错误
+   * 执行 定义的单个 job
    * */
-  async runTest(){
-    await this._prepareEnv();
-    const testJobs = this.workflowInfo?.test || [];
-
-    for(let i=0; i<testJobs.length; i++){
-      await this.runJob(testJobs[i],`test${Date.now()}`)
+  async _runJob(job: Job,jobsContextId: string){
+    if(!job) return;
+    let jobResponse;
+    job._state = TaskState.running;
+    /**
+     * 如果 job 是一个 matrix 任务，则要进行叉乘创建子 job。
+     */
+    const variables = this._getRuntime(jobsContextId) || {};
+    // todo 仅替换 job ，不处理 step，step 运行时再替换，避免初始化时 step 需要的变量还未生成
+    job = replaceTemplates<Job>(job,variables)
+    let markSuccess = true;
+    // todo runjob 支持if 判断
+    // todo jobs 之间取值？
+    if(job.strategy?.matrix){
+      const matrixVariables = generateMatrixTasks(job.strategy.matrix) || [];
+      for (let i=0; i<matrixVariables.length; i++){
+        for(let j=0; j<matrixVariables[i].length; j++){
+          const contextId = (job.id || job.name) + 'matrix'+i+j;
+          const stepVariable = matrixVariables[i][j];
+          this._setRuntime('matrix',stepVariable,contextId)
+          // 每一个 matrix 相互不受影响运行
+          try{
+            jobResponse = await this._runSteps(job.steps,contextId);
+            this._setRuntime(`jobs.${job.id||job.name}.outputs`,jobResponse,jobsContextId)
+          }catch (e) {
+            markSuccess = false;
+          }
+        }
+      }
+    } else {
+      const {id='',name='',} = job || {};
+      const contextId = 'job_'+id + name // + Date.now();
+      jobResponse = await this._runSteps(job.steps,contextId)
+      this._setRuntime(`jobs.${id||name}.outputs`,jobResponse,jobsContextId)
     }
 
-    return testJobs.every(function (job) {
-      return job._state !== TaskState.fail
-    })
+    job._state = markSuccess ? TaskState.complete : TaskState.fail;
+    return jobResponse;
   }
 
-  check(): Promise<boolean> {
-    return Promise.resolve(false);
+  /**
+   * 执行多个任务组；用于控制并发、串行、失败控制等
+   * */
+  async _runSteps(steps: Step[],contextId:string){
+    let response;
+    for (let i = 0; i < steps.length; i++) {
+      try {
+        // steps 集合的最终返回，取决于最后一个任务的执行结果
+        response = await this._runStep(steps[i],contextId);
+      }catch (e) {
+        if(steps[i]['continue-on-error']){
+          console.warn('允许失败，继续执行',e)
+        }else{
+          throw e;
+        }
+      }
+    }
+    return response;
+  }
+
+  /**
+   * 执行单个 step 任务
+   * */
+  async _runStep(step: Step,stepsContextId:string){
+    if(!step){
+      return
+    }
+    const { uses='' } = step;
+
+    step._state = TaskState.running;
+
+    const variables = this._getContextVariables(stepsContextId)
+
+
+    if(step.debug){
+      console.debug('debug',step)
+      debugger
+    }
+    let response = null;
+
+    /**条件检测**/
+    const ifCheckResult = step.if ? ifCheck(step.if,variables) : true;
+    if(!ifCheckResult){
+      step._state = TaskState.skip;
+    } else{
+      const withArgs = step.with ? replaceTemplates(step.with,variables) : undefined
+      try{
+        if(this.option.hooks?.beforeStep){
+          this.option.hooks?.beforeStep(step,withArgs);
+        }
+        // 由用户自行提供 action or 系统默认 actions
+        const action = await this.option.registerAction(uses || '') || await getAction(uses);
+        if(uses && !action){
+          step._state = TaskState.fail;
+          throw Error('without action::'+ uses)
+        }
+
+        if(step.run && step.run.length){
+          step.run.forEach(function (terminate) {
+            response = exprEval(terminate,variables)
+          })
+        }else{
+          try{
+            response = action ? await action(withArgs) : withArgs;
+          }catch (e) {
+            step._state = TaskState.fail;
+            throw e;
+          }
+        }
+
+        this._setRuntime(`steps.${step.id || step.name}.outputs`, response,stepsContextId)
+      }catch (e) {
+        step._state = TaskState.fail;
+        throw e;
+      }
+      step._state = TaskState.complete;
+      if(this.option.hooks?.afterStep){
+        this.option.hooks?.afterStep(step,response);
+      }
+    }
+
+    /**由用户自行控制step 退出码*/
+    let exitCode: unknown = 0;
+    if(step.exit){
+      if(typeof step.exit === 'number'){
+        exitCode = step.exit;
+      }else{
+        exitCode = exprEval(step.exit,this._getContextVariables(stepsContextId))
+      }
+
+      if(exitCode!==0){
+        step._state = TaskState.fail;
+        throw Error((step.id || step.name) + ' exit code is not equal 0 :' + exitCode);
+      }
+    }
+    return response;
   }
 }